merge/resolve conflicts (some more to do after this one)

+++ gnu/usr.bin/perl/ext/SDBM_File/sdbm/pair.c
@@ -281,7 +281,7 @@ chkpage(char *pag)
 	register int off;
 	register short *ino = (short *) pag;
 
-	if ((n = ino[0]) < 0 || n > PBLKSIZ / sizeof(short))
+	if ((n = ino[0]) < 0 || n > (int)(PBLKSIZ / sizeof(short)))
 		return 0;
 
 	if (n > 0) {
